+++ android-sdk-couriers/src/main/java/com/citibox/courier/sdk/deeplink/DeepLinkRetrievalContract.kt
@@ -5,34 +5,26 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import androidx.activity.result.contract.ActivityResultContract
-import com.citibox.courier.sdk.domain.DeliveryParams
-import com.citibox.courier.sdk.domain.DeliveryResult
+import com.citibox.courier.sdk.domain.RetrievalParams
+import com.citibox.courier.sdk.domain.RetrievalResult
 import com.citibox.courier.sdk.domain.TransactionCancel
 import com.citibox.courier.sdk.domain.TransactionError
 import com.citibox.courier.sdk.domain.TransactionResult
-import java.security.MessageDigest
 
-internal class DeepLinkDeliveryContract :
-    ActivityResultContract<DeliveryParams, DeliveryResult>() {
-    override fun createIntent(context: Context, input: DeliveryParams): Intent {
+internal class DeepLinkRetrievalContract :
+    ActivityResultContract<RetrievalParams, RetrievalResult>() {
+    override fun createIntent(context: Context, input: RetrievalParams): Intent {
         val uriBuilder = Uri.Builder()
             .scheme(SCHEME)
             .authority(AUTHORITY)
             .appendPath(PATH)
             .appendQueryParameter(QUERY_ACCESS_TOKEN, input.accessToken.trim())
-            .appendQueryParameter(QUERY_TRACKING, input.tracking.trim())
-            .appendQueryParameter(QUERY_DIMENSIONS, input.dimensions?.trim().orEmpty())
-
-        if (input.isPhoneHashed) {
-            uriBuilder.appendQueryParameter(QUERY_HASH, input.recipientPhone.calculateSHA256())
-        } else {
-            uriBuilder.appendQueryParameter(QUERY_PHONE, input.recipientPhone)
-        }
+            .appendQueryParameter(QUERY_CITIBOX_ID, input.citiboxId.trim())
 
         return Intent(Intent.ACTION_VIEW).setData(uriBuilder.build())
     }
 
-    override fun parseResult(resultCode: Int, intent: Intent?): DeliveryResult =
+    override fun parseResult(resultCode: Int, intent: Intent?): RetrievalResult =
         when (resultCode) {
             Activity.RESULT_OK -> runCatching {
                 val validIntent = requireNotNull(intent) {
@@ -44,14 +36,12 @@ internal class DeepLinkDeliveryContract :
                 val citiboxId = requireNotNull(validIntent.extras?.getInt(EXTRA_CITIBOX_ID)) {
                     "Missing Citibox ID"
                 }
-                val deliveryId = validIntent.getStringExtra(EXTRA_DELIVERY_ID).orEmpty()
-                DeliveryResult.Success(
+                RetrievalResult.Success(
                     boxNumber = boxNumber,
                     citiboxId = citiboxId,
-                    deliveryId = deliveryId,
                 )
             }.getOrElse {
-                DeliveryResult.Error(TransactionError.DATA_NOT_RECEIVED.code)
+                RetrievalResult.Error(TransactionError.DATA_NOT_RECEIVED.code)
             }
 
             else ->
@@ -65,45 +55,32 @@ internal class DeepLinkDeliveryContract :
                     intent?.hasExtra(TransactionResult.FAILURE_CODE_KEY.code) == true ->
                         parseFailure(intent = intent)
 
-                    else -> DeliveryResult.Error(TransactionCancel.OTHER.code)
+                    else -> RetrievalResult.Error(TransactionCancel.OTHER.code)
                 }
         }
 
-    private fun parseError(intent: Intent): DeliveryResult {
+    private fun parseError(intent: Intent): RetrievalResult {
         val typeError =
             requireNotNull(intent.getStringExtra(TransactionResult.ERROR_CODE_KEY.code)) {
                 "Missing type error"
             }
-        return DeliveryResult.Error(typeError)
+        return RetrievalResult.Error(typeError)
     }
 
-    private fun parseCancel(intent: Intent): DeliveryResult {
+    private fun parseCancel(intent: Intent): RetrievalResult {
         val typeError =
             requireNotNull(intent.getStringExtra(TransactionResult.CANCEL_CODE_KEY.code)) {
                 "Missing type cancellation"
             }
-        return DeliveryResult.Cancel(typeError)
+        return RetrievalResult.Cancel(typeError)
     }
 
-    private fun parseFailure(intent: Intent): DeliveryResult {
+    private fun parseFailure(intent: Intent): RetrievalResult {
         val typeError =
             requireNotNull(intent.getStringExtra(TransactionResult.FAILURE_CODE_KEY.code)) {
                 "Missing type failure"
             }
-        return DeliveryResult.Failure(typeError)
-    }
-
-    private fun String.calculateSHA256(): String {
-        val digest = MessageDigest.getInstance("SHA-256")
-        val data = digest.digest(this.toByteArray())
-
-        return bin2hex(data)
-    }
-
-    private fun bin2hex(data: ByteArray): String {
-        val hex = StringBuilder(data.size * 2)
-        for (b in data) hex.append(String.format("%02x", b.toInt() and 0xFF))
-        return hex.toString()
+        return RetrievalResult.Failure(typeError)
     }
 
     companion object {
@@ -111,13 +88,9 @@ internal class DeepLinkDeliveryContract :
         private const val AUTHORITY = "couriers.citibox.com"
         private const val PATH = "transaction"
         private const val QUERY_ACCESS_TOKEN = "access_token"
-        private const val QUERY_TRACKING = "tracking"
-        private const val QUERY_PHONE = "recipient_phone"
-        private const val QUERY_HASH = "recipient_hash"
-        private const val QUERY_DIMENSIONS = "dimensions"
+        private const val QUERY_CITIBOX_ID = "citibox_id"
 
         private const val EXTRA_BOX_NUMBER = "box_number"
         private const val EXTRA_CITIBOX_ID = "citibox_id"
-        private const val EXTRA_DELIVERY_ID = "delivery_id"
     }
 }