+++ app/src/main/java/com/citibox/courier/sdk/example/compose/MainLayout.kt
@@ -8,10 +8,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Button
-import androidx.compose.material3.Card
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
@@ -25,6 +23,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import com.citibox.courier.sdk.example.models.MainState
 import com.citibox.courier.sdk.example.ui.theme.components.InputDataField
+import com.citibox.courier.sdk.webview.models.WebAppEnvironment
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -35,6 +34,7 @@ fun MainLayout(
     onPhoneChanged: (String) -> Unit,
     onPhoneHashedChanged: (Boolean) -> Unit,
     onDimensionsChanged: (String) -> Unit,
+    onEnvironmentChanged: (WebAppEnvironment) -> Unit,
     onLaunchClicked: () -> Unit,
 ) {
     Scaffold(
@@ -96,6 +96,10 @@ fun MainLayout(
                 label = "Dimensions ([mm]x[mm]x[mm])",
                 onValueChanged = onDimensionsChanged,
             )
+            EnvironmentSelector(
+                actual = state.environment,
+                onItemClicked = onEnvironmentChanged
+            )
             Button(
                 modifier = Modifier
                     .fillMaxWidth()
@@ -107,44 +111,6 @@ fun MainLayout(
             ) {
                 Text(
                     text = "Launch Citibox Delivery",
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun ResultCard(resultMessage: String) {
-    Card(
-        modifier = Modifier
-            .padding(
-                horizontal = 16.dp,
-                vertical = 8.dp,
-            )
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(
-                    vertical = 16.dp,
-                    horizontal = 0.dp
-                )
-        ) {
-            Text(
-                text = "Result", style = MaterialTheme.typography.titleLarge,
-                modifier = Modifier
-                    .padding(horizontal = 16.dp)
-            )
-
-            SelectionContainer {
-                Text(
-                    text = resultMessage,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(
-                            top = 8.dp,
-                            start = 16.dp,
-                            end = 16.dp,
-                        )
                 )
             }
         }