+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
+## [1.1.0] 2023-12-14
+### Added
+- Environments from selector
+- Added deeplink contract for retrieval
+
 ## [1.0.0] 2023-12-14
 ### Changed
 - When the web view receives an error then bounces back to the SDK 
-- Added web view contract
-- Added deeplink contract
+- Added web view contract for delivery
+- Added deeplink contract for delivery